# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Seat Reservation API",
    "description": "\nSells assigned seats for a show and guarantees each seat is sold exactly once, even when\nthousands of buyers want the same seat in the same second.\n\n**Try it from this page**\n\n1. `POST /auth/token` with `{\"user_id\": \"alice\"}`, then copy `access_token`.\n2. Click **Authorize**: paste the token under **BearerAuth**, and the admin key under\n   **AdminKey** (only needed to create shows and run the audit).\n3. `POST /shows` to create a show, then `POST /shows/{show_id}/reserve`.\n4. `GET /shows/{show_id}` shows every seat: `available + held + confirmed == total` always holds.\n\n**Rules worth knowing**\n\n- Identity comes only from the token; a `user_id` in a request body is ignored.\n- Money is integer paise; a float or string price is rejected.\n- A multi-seat request is all-or-nothing.\n- The same idempotency key returns the same reservation (`200` replay); the same key with\n  other seats is `409`.\n- Every error has the same shape, `{\"error\": {\"code\", \"message\", ...}, \"request_id\"}`.\n  Declines are `409`, never `5xx`. If the database is unreachable: `503` with `Retry-After`.\n\nSource, write-up and the one-command burst:\n[github.com/jaymovaliya26/seat-reservation](https://github.com/jaymovaliya26/seat-reservation)\n",
    "version": "1.1.0"
  },
  "paths": {
...
    "/shows": {
      "post": {
        "tags": [
          "shows"
        ],
        "summary": "Create a show (admin)",
        "operationId": "create_show_shows_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShowRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShowOut"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token (or missing admin key)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                },
                "examples": {
                  "missing token": {
                    "value": {
                      "error": {
                        "code": "unauthorized",
                        "message": "Missing bearer token"
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  },
                  "expired token": {
                    "value": {
                      "error": {
                        "code": "unauthorized",
                        "message": "Token is invalid or expired"
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Wrong admin key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                },
                "examples": {
                  "wrong key": {
                    "value": {
                      "error": {
                        "code": "forbidden",
                        "message": "Invalid admin key"
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "The request is malformed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                },
                "examples": {
                  "invalid body": {
                    "value": {
                      "error": {
                        "code": "invalid_request",
                        "message": "The request is not valid",
                        "problems": [
                          {
                            "field": "body.price_paise",
                            "problem": "Input should be a valid integer"
                          }
                        ]
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  },
                  "unknown seats": {
                    "value": {
                      "error": {
                        "code": "unknown_seats",
                        "message": "Seats do not exist in this show: Z9",
                        "seats": [
                          "Z9"
                        ]
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "The database is unreachable or overloaded; retry with the same key after `Retry-After` seconds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                },
                "examples": {
                  "unavailable": {
                    "value": {
                      "error": {
                        "code": "database_unavailable",
                        "message": "The database is unreachable; retry shortly with the same idempotency key"
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  },
                  "busy": {
                    "value": {
                      "error": {
                        "code": "database_busy",
                        "message": "The database is overloaded; retry shortly with the same idempotency key"
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "AdminKey": []
          }
        ]
      }
    },
    "/shows/{show_id}": {
      "get": {
        "tags": [
          "shows"
        ],
        "summary": "Read a show's seat map",
        "operationId": "get_show_shows__show_id__get",
        "parameters": [
          {
            "name": "show_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Show Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShowOut"
                }
              }
            }
          },
          "404": {
            "description": "Not found (including another user's reservation)",
            "content": {
              "application/json": {
                "examples": {
                  "not found": {
                    "value": {
                      "error": {
                        "code": "not_found",
                        "message": "Show not found"
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "The request is malformed",
            "content": {
              "application/json": {
                "examples": {
                  "invalid body": {
                    "value": {
                      "error": {
                        "code": "invalid_request",
                        "message": "The request is not valid",
                        "problems": [
                          {
                            "field": "body.price_paise",
                            "problem": "Input should be a valid integer"
                          }
                        ]
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  },
                  "unknown seats": {
                    "value": {
                      "error": {
                        "code": "unknown_seats",
                        "message": "Seats do not exist in this show: Z9",
                        "seats": [
                          "Z9"
                        ]
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "The database is unreachable or overloaded; retry with the same key after `Retry-After` seconds",
            "content": {
              "application/json": {
                "examples": {
                  "unavailable": {
                    "value": {
                      "error": {
                        "code": "database_unavailable",
                        "message": "The database is unreachable; retry shortly with the same idempotency key"
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  },
                  "busy": {
                    "value": {
                      "error": {
                        "code": "database_busy",
                        "message": "The database is overloaded; retry shortly with the same idempotency key"
                      },
                      "request_id": "9f2c4a1e0b7d4c55a3e1f0d2b6c8e4a7"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateShowRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "seats": {
            "items": {
              "type": "string",
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,15}$"
            },
            "type": "array",
            "maxItems": 50000,
            "minItems": 1,
            "title": "Seats"
          },
          "price_paise": {
            "type": "integer",
            "maximum": 1000000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Price Paise"
          },
          "per_user_limit": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Per User Limit"
          }
        },
        "type": "object",
        "required": [
          "name",
          "seats",
          "price_paise"
        ],
        "title": "CreateShowRequest",
        "examples": [
          {
            "name": "friday-night",
            "per_user_limit": 4,
            "price_paise": 25000,
            "seats": [
              "A1",
              "A2",
              "A3",
              "A11",
              "A12",
              "A13"
            ]
          }
        ]
      },
      "ErrorBody": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorInfo"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "error",
          "request_id"
        ],
        "title": "ErrorBody"
      },
      "ErrorInfo": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorInfo"
      },
...
      "SeatCounts": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "available": {
            "type": "integer",
            "title": "Available"
          },
          "held": {
            "type": "integer",
            "title": "Held"
          },
          "confirmed": {
            "type": "integer",
            "title": "Confirmed"
          }
        },
        "type": "object",
        "required": [
          "total",
          "available",
          "held",
          "confirmed"
        ],
        "title": "SeatCounts"
      },
      "SeatOut": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "status": {
            "type": "string",
            "enum": [
              "available",
              "held",
              "confirmed"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "label",
          "status"
        ],
        "title": "SeatOut"
      },
      "ShowOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "price_paise": {
            "type": "integer",
            "title": "Price Paise"
          },
          "per_user_limit": {
            "type": "integer",
            "title": "Per User Limit"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "counts": {
            "$ref": "#/components/schemas/SeatCounts"
          },
          "seats": {
            "items": {
              "$ref": "#/components/schemas/SeatOut"
            },
            "type": "array",
            "title": "Seats"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "price_paise",
          "per_user_limit",
          "created_at",
          "counts",
          "seats"
        ],
        "title": "ShowOut"
      },
...
    },
    "securitySchemes": {
      "AdminKey": {
        "type": "apiKey",
        "description": "Admin key, needed only to create shows and run the reconcile audit.",
        "in": "header",
        "name": "X-Admin-Key"
      },
      "BearerAuth": {
        "type": "http",
        "description": "A token from POST /auth/token. Paste only the token; Swagger adds 'Bearer '.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}